+++ scripts/clean.sh
@@ -6,6 +6,8 @@ PIGEN_DIR=${DEPS_DIR}/pi-gen
 function clean_setup {
   git submodule update --init --recursive
   pushd ${PIGEN_DIR}
+  rm */SKIP
+  rm config
   git reset --hard HEAD
   git clean -f -d
   popd